Velocity, attribute, and every metre in the same unit
nuScenes scores five errors and a box carries three of them. The other
two are now emitted: a velocity vector, from a central difference of the
track's world position rotated into the ego frame, and an attribute from
the same 1 m/s verdict task 06 already makes. `parked` and `stopped` stay
merged -- both are stationary and which one it is lives in the driver's
intent, not in any sensor.

    automobile (4.1m, 6.8m, 0.5m) size 4.3mx1.9mx1.5m heading 6
      vel +0.3m/s -0.0m/s stopped

Metres are one unit again. Fusing the suffix caught `37m` in a range but
not the `4.1` in a coordinate, so the loss saw two metre buckets and
would not compare a range of 40 m against an x of 40 m -- the same
quantity. Coordinates, extents and waypoints carry the suffix now, and
what is left bare is what genuinely is not a length: counts, heading
sectors, track ids.

+++ training/radar_encoder.py
@@ -279,7 +279,9 @@ class Block(nn.Module):
         mask = None
         if attn_bias is not None:
             heads = self.attn.num_heads
-            mask = attn_bias.repeat_interleave(heads, dim=0)
+            # The bias has to carry the query's dtype -- the model runs in
+            # bfloat16 and a float32 mask is rejected outright.
+            mask = attn_bias.repeat_interleave(heads, dim=0).to(q.dtype)
         attended, _ = self.attn(q, kv, kv, need_weights=False,
                                 key_padding_mask=key_padding_mask,
                                 attn_mask=mask)
